traefik-sni-hub/panel/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
PyJWT==2.10.1
bcrypt==4.2.1
toml==0.10.2
docker==7.1.0
